backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""

readme = "../README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "owslib (>=0.34.1,<0.35.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "pyproj (>=3.7.1,<4.0.0)",
    "google-genai (>=1.25.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "cdsapi (>=0.7.6,<0.8.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-core (>=0.3.72,<0.4.0)",
    "langchain-tavily (>=0.2.10,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "deep-translator (>=1.11.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "google-cloud-translate (>=3.21.1,<4.0.0)",
    "google-cloud-speech (>=2.28.0,<3.0.0)",
    "google-cloud-texttospeech (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
]

[tool.poetry]
packages = [   
    { include = "auth", from = "src" },
    { include = "ext_apis", from = "src" },
    { include = "db", from ="src"},
    {include = "models", from = "src"},
    {include = "routes", from = "src"}
    
]


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.35.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
